requirements.txt
pymongo
pandas
numpy
scikit-learn
pyyaml
mlflow
dagshub
python-dotenv
joblib
scipy